# ffs-api/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "description": "API for user authentication, creation, and management of schools, districts, bids, and tasks"
  },
  "paths": {
...
    "/api/organizations/invite": {
      "post": {
        "summary": "Invite a new organization",
        "description": "Invite a new organization to the platform with specified type and name.",
        "tags": ["Organizations"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address for the organization invitation"
                  },
                  "organization_type": {
                    "type": "string",
                    "enum": ["Coop", "Single-District"],
                    "description": "Type of organization"
                  },
                  "name": {
                    "type": "string",
                    "description": "Name of the organization"
                  }
                },
                "required": ["email", "organization_type", "name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization invitation sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not Super-Admin",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/invite/accept": {
      "post": {
        "summary": "Accept an organization invitation and register user",
        "description": "Accept an organization invitation and register a new user with the provided details.",
        "tags": ["Organizations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "description": "User's password"
                  },
                  "first_name": {
                    "type": "string",
                    "description": "User's first name"
                  },
                  "last_name": {
                    "type": "string",
                    "description": "User's last name"
                  },
                  "token": {
                    "type": "string",
                    "description": "Invitation token"
                  }
                },
                "required": ["email", "password", "first_name", "last_name", "token"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization invitation accepted and user successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    },
                    "user_id": {
                      "type": "integer",
                      "description": "ID of the registered user"
                    },
                    "organization_id": {
                      "type": "integer",
                      "description": "ID of the organization"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token, or invalid email",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{id}": {
      "put": {
        "summary": "Update organization",
        "description": "Update an organization's details, such as address, phone, or email.",
        "tags": ["Organizations"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the organization to update",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid organization ID or request body",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdateOrganizationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the organization"
          },
          "address": {
            "type": "string",
            "description": "Physical address of the organization"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number of the organization"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Contact email of the organization"
          }
        }
      },
...
    }
  }
}